+++ src/App.js
@@ -1,10 +1,6 @@
 import React from "react";
-
-const todoList = [
-  { id: 1, title: "Learn React" },
-  { id: 2, title: "Review JavaScript" },
-  { id: 3, title: "Complete assignment" },
-];
+import AddTodoForm from "./AddTodoForm";
+import TodoList from "./TodoList";
 
 function App() {
   return (
@@ -12,16 +8,8 @@ function App() {
       <header>
         <h1>Todo list</h1>
       </header>
-      <ul>
-        {/* render the TodoList */}
-        {todoList.map((item) => {
-          return (
-            <li>
-              {item.id} {item.title}
-            </li>
-          );
-        })}
-      </ul>
+      <AddTodoForm />
+      <TodoList />
     </div>
   );
 }
